给删除联系人添加判断

+++ src/main/java/com/bodejidi/AdministratorContactServlet.java
@@ -165,55 +165,58 @@ public class AdministratorContactServlet extends HttpServlet{
         throws IOException,ServletException{
             
         String sql;
-        int result;
+        String action = request.getParameter("action");
         Connection connection = null;
         Statement statement = null;
         ResultSet resultSet = null;
         
-        try{
-            Class.forName("com.mysql.jdbc.Driver").newInstance();
-        }catch(Exception ex){
-            //handl the error
-        }
-        if(request.getParameter("contactId") != null && request.getParameter("contactId") != ""){
+        if(action.equals("Delete")){
             try{
-                sql = "delete from contact where id=" 
-                    + request.getParameter("contactId");
-                
-                Contact contact = new Contact();
-                
-                connection = DriverManager.getConnection("jdbc:mysql://localhost/test?user=root&password=");
-                statement = connection.createStatement();
-                result = statement.executeUpdate(sql); 
-                response.getWriter().println("The Contact has been delete");
-           
-            }catch(SQLException sqle){
-                response.getWriter().println("can not connect to DB");
-                response.getWriter().println(sqle.getMessage());
-                sqle.printStackTrace();
+                Class.forName("com.mysql.jdbc.Driver").newInstance();
+            }catch(Exception ex){
+                //handl the error
             }
             
-            if(resultSet != null){
+            if(request.getParameter("contactId") != null && request.getParameter("contactId") != ""){
                 try{
-                    resultSet.close();
-                }catch(Exception ex){
-                
+                    sql = "delete from contact where id=" 
+                        + request.getParameter("contactId");
+                    
+                    Contact contact = new Contact();
+                    
+                    connection = DriverManager.getConnection("jdbc:mysql://localhost/test?user=root&password=");
+                    statement = connection.createStatement();
+                    statement.executeUpdate(sql); 
+                    response.getWriter().println("The Contact has been delete");
+               
+                }catch(SQLException sqle){
+                    response.getWriter().println("can not connect to DB");
+                    response.getWriter().println(sqle.getMessage());
+                    sqle.printStackTrace();
                 }
-            }
-            
-            if(statement != null){
-                try{
-                    statement.close();
-                }catch(Exception ex){
                 
+                if(resultSet != null){
+                    try{
+                        resultSet.close();
+                    }catch(Exception ex){
+                    
+                    }
                 }
-            }
-            
-            if(connection != null){
-                try{
-                    connection.close();
-                }catch(Exception e){
                 
+                if(statement != null){
+                    try{
+                        statement.close();
+                    }catch(Exception ex){
+                    
+                    }
+                }
+                
+                if(connection != null){
+                    try{
+                        connection.close();
+                    }catch(Exception e){
+                    
+                    }
                 }
             }
         }
